+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 * nostr: add `with_capacity`, `from_list`, `from_text` and `parse` constructors to `Tags` struct ([Yuki Kishimoto])
 * nostr: add `Tags::dedup` method ([Yuki Kishimoto])
 * nostr: add `EncryptedSecretKey::decrypt` method ([Yuki Kishimoto])
+* nostr: add `Nip19Coordinate` struct ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
 * ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])